# docs/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "AI Agent Productivity Platform - Open API",
        "contact": {},
        "version": "v0.0.1"
    },
    "basePath": "/openapi/v1",
    "paths": {
...
        "/knowledge/doc": {
            "delete": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "删除文档",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openapi"
                ],
                "summary": "删除文档",
                "parameters": [
                    {
                        "description": "删除文档请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.DeleteDocReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/knowledge/doc/config": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "获取文档配置信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openapi"
                ],
                "summary": "获取文档配置信息",
                "parameters": [
                    {
                        "type": "string",
                        "name": "docId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "knowledgeId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.DocConfigResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/knowledge/doc/export": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "知识库文档导出",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openapi"
                ],
                "summary": "知识库文档导出",
                "parameters": [
                    {
                        "description": "知识库文档导出请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.KnowledgeDocExportReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/knowledge/doc/import": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "上传文档",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openapi"
                ],
                "summary": "上传文档",
                "parameters": [
                    {
                        "description": "文档上传请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.DocImportReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/knowledge/doc/import/tip": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "获取知识库异步上传任务提示：有正在执行的异步上传任务/最近一次上传任务的失败信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openapi"
                ],
                "summary": "获取知识库异步上传任务提示",
                "parameters": [
                    {
                        "type": "string",
                        "name": "knowledgeId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/knowledge/doc/list": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "获取知识库文档列表，不展示状态为无效（-1）的文档数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openapi"
                ],
                "summary": "获取文档列表",
                "parameters": [
                    {
                        "type": "string",
                        "name": "docName",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "图谱状态：-1.全部 0.待处理 1.解析中 2.解析成功 3.解析失败",
                        "name": "graphStatus",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "knowledgeId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "metaValue",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "pageNo",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "文档状态：-1-全部， 0-待处理， 1- 处理完成， 2-正在审核中，3-正在解析中，4-审核未通过，5-解析失败",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.DocPageResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/knowledge/doc/update/config": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "更新文档配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openapi"
                ],
                "summary": "更新文档配置",
                "parameters": [
                    {
                        "description": "更新文档配置请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.DocConfigUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.DeleteDocReq": {
            "type": "object",
            "required": [
                "docIdList",
                "knowledgeId"
            ],
            "properties": {
                "docIdList": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "knowledgeId": {
                    "type": "string"
                }
            }
        },
...
        "request.DocConfigUpdateReq": {
            "type": "object",
            "required": [
                "docAnalyzer",
                "docIdList",
                "docSegment",
                "knowledgeId"
            ],
            "properties": {
                "docAnalyzer": {
                    "description": "文档解析类型 text / ocr  / model",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "docIdList": {
                    "description": "文档id列表，当更新配置时需要传",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "docImportType": {
                    "description": "文档导入类型，0：文件上传，1：单条url上传，2.文件url上传",
                    "type": "integer"
                },
                "docMetaData": {
                    "description": "元数据",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.DocMetaData"
                    }
                },
                "docPreprocess": {
                    "description": "文本预处理规则 replaceSymbols / deleteLinks",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "docSegment": {
                    "description": "文档分段配置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/request.DocSegment"
                        }
                    ]
                },
                "knowledgeId": {
                    "description": "知识库id",
                    "type": "string"
                },
                "parserModelId": {
                    "description": "模型解析或ocr模型id",
                    "type": "string"
                }
            }
        },
        "request.DocImportReq": {
            "type": "object",
            "required": [
                "docAnalyzer",
                "docInfoList",
                "docSegment",
                "knowledgeId"
            ],
            "properties": {
                "docAnalyzer": {
                    "description": "文档解析类型 text / ocr  / model",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "docImportType": {
                    "description": "文档导入类型，0：文件上传，1：单条url上传，2.文件url上传",
                    "type": "integer"
                },
                "docInfoList": {
                    "description": "上传文档列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.DocInfo"
                    }
                },
                "docMetaData": {
                    "description": "元数据",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.DocMetaData"
                    }
                },
                "docPreprocess": {
                    "description": "文本预处理规则 replaceSymbols / deleteLinks",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "docSegment": {
                    "description": "文档分段配置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/request.DocSegment"
                        }
                    ]
                },
                "knowledgeId": {
                    "description": "知识库id",
                    "type": "string"
                },
                "parserModelId": {
                    "description": "模型解析或ocr模型id",
                    "type": "string"
                }
            }
        },
        "request.DocInfo": {
            "type": "object",
            "properties": {
                "docId": {
                    "description": "文档id",
                    "type": "string"
                },
                "docName": {
                    "description": "文档名称",
                    "type": "string"
                },
                "docSize": {
                    "description": "文档类型",
                    "type": "integer"
                },
                "docType": {
                    "description": "文档类型",
                    "type": "string"
                },
                "docUrl": {
                    "description": "文档url",
                    "type": "string"
                }
            }
        },
        "request.DocMetaData": {
            "type": "object",
            "properties": {
                "metaId": {
                    "description": "元数据id",
                    "type": "string"
                },
                "metaKey": {
                    "description": "key",
                    "type": "string"
                },
                "metaRule": {
                    "description": "正则表达式",
                    "type": "string"
                },
                "metaValue": {
                    "description": "确定值",
                    "type": "string"
                },
                "metaValueType": {
                    "description": "string，number，time",
                    "type": "string"
                },
                "option": {
                    "description": "option:add(新增)、update(更新)、delete(删除),update 和delete 的时候metaId 不能为空",
                    "type": "string"
                }
            }
        },
...
        "request.KnowledgeDocExportReq": {
            "type": "object",
            "required": [
                "knowledgeId"
            ],
            "properties": {
                "docIdList": {
                    "description": "文档id",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "knowledgeId": {
                    "description": "知识库id",
                    "type": "string"
                }
            }
        },
...
        "response.DocConfigResult": {
            "type": "object",
            "properties": {
                "docAnalyzer": {
                    "description": "文档解析类型",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "docImportType": {
                    "description": "文档导入类型，0：文件上传，1：url上传，2.批量url上传",
                    "type": "integer"
                },
                "docPreprocess": {
                    "description": "文本预处理规则",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "docSegment": {
                    "description": "分段信息配置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/response.DocSegment"
                        }
                    ]
                },
                "parserModelId": {
                    "description": "ocr模型id",
                    "type": "string"
                }
            }
        },
        "response.DocKnowledgeInfo": {
            "type": "object",
            "properties": {
                "graphSwitch": {
                    "type": "integer"
                },
                "knowledgeId": {
                    "type": "string"
                },
                "knowledgeName": {
                    "type": "string"
                },
                "showGraphReport": {
                    "type": "boolean"
                }
            }
        },
        "response.DocPageResult": {
            "type": "object",
            "properties": {
                "docKnowledgeInfo": {
                    "$ref": "#/definitions/response.DocKnowledgeInfo"
                },
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.ListDocResp"
                    }
                },
                "pageNo": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "response.DocSegment": {
            "type": "object",
            "required": [
                "segmentMethod"
            ],
            "properties": {
                "maxSplitter": {
                    "description": "可分隔最大值（只有自定义分段必填）",
                    "type": "integer"
                },
                "overlap": {
                    "description": "可重叠值（只有自定义分段必填）",
                    "type": "number"
                },
                "segmentMethod": {
                    "description": "分段方法 0：通用分段；1：父子分段",
                    "type": "string"
                },
                "segmentType": {
                    "description": "分段方式，只有通用分段必填 0：自动分段；1：自定义分段",
                    "type": "string"
                },
                "splitter": {
                    "description": "分隔符（只有自定义分段必填）",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "subMaxSplitter": {
                    "description": "可分隔最大值（只有父子分段必填）",
                    "type": "integer"
                },
                "subSplitter": {
                    "description": "分隔符（只有父子分段必填）",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    }
}